app/worker/requirements.txt
boto3==1.38.0
prometheus_client==0.21.1
pytesseract==0.3.13
Pillow==11.2.1
